pyproject.toml
[tool.poetry]
name = "psychoanalyze"
version = "0.1.0"
description = "Tools for handling, analyzing, and visualizing psychophysics data."
authors = ["Tyler Schlichenmeyer <tyler.schlichenmeyer@gmail.com>"]
license = "MIT"
readme = "README.rst"


[tool.poetry.dependencies]
python = "^3.9.1"
plotly = "^4.14.3"
pandas = "^1.2.4"
pandera = {extras = ["all"], version = "^0.6.3"}
hypothesis = "^6.9.0"
datatest = "^0.11.1"
dash = "^1.20.0"
jupyterlab = "^3.0.14"
psignifit = {git = "https://github.com/wichmann-lab/python-psignifit.git"}
tables = "^3.6.1"
numpy = "^1.20.2"
ipywidgets = ">=7.5"
dash-daq = "^0.5.0"
tqdm = "^4.60.0"
statsmodels = "^0.12.2"

[tool.poetry.dev-dependencies]
pytest = "^6.2.3"
black = {version = "^20.8b1", allow-prereleases = true}
flake8 = "^3.9.0"
nb-clean = "^2.0.2"
pytest-sugar = "^0.9.4"
factory-boy = "^3.2.0"
mimesis = "^4.1.3"
mimesis-factory = "^1.2.0"
pytest-factoryboy = "^2.1.0"
dash = {extras = ["testing"], version = "^1.20.0"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = ["tests"]